LPS-55719 moves java
