+++ modules/apps/roles-admin/roles-admin-web/src/com/liferay/roles/admin/web/customattributes/RoleCustomAttributesDisplay.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portlet.rolesadmin;
+package com.liferay.roles.admin.web.customattributes;
 
 import com.liferay.portal.model.Role;
 import com.liferay.portal.theme.ThemeDisplay;
